AICDAP/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
resend==0.6.0
python-dotenv==1.0.0
supabase==2.0.0
python-multipart==0.0.6
jinja2==3.1.2
httpx>=0.24.0,<0.25.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

pandas==2.2.2
numpy==1.26.4
scikit-learn==1.3.2
joblib==1.3.2

nltk==3.8.1
tldextract==3.6.0

pillow==10.0.1
opencv-python==4.8.1.78
pytesseract==0.3.10

networkx==3.2.1

# ======================
# TORCH (CPU ONLY SAFE BUILD)
# ======================
torch==2.2.2+cpu ; platform_system=="Windows"
torchvision==0.17.2+cpu ; platform_system=="Windows"
torchaudio==2.2.2+cpu ; platform_system=="Windows"

--extra-index-url https://download.pytorch.org/whl/cpu

# Graph ML
torch-geometric==2.4.0